chore: move assert package to new internal/ dir
